Cor auth

+++ src/main/java/ru/zubov/pizzacloud/controller/rest/IngredientController.java
@@ -2,12 +2,15 @@ package ru.zubov.pizzacloud.controller.rest;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
-import org.springframework.security.access.annotation.Secured;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.*;
 import ru.zubov.pizzacloud.entity.Ingredient;
+import ru.zubov.pizzacloud.entity.SignUpDto;
 import ru.zubov.pizzacloud.entity.dtos.IngredientDto;
 import ru.zubov.pizzacloud.entity.mapper.IngredientMapper;
 import ru.zubov.pizzacloud.repository.IngredientRepository;
+import ru.zubov.pizzacloud.service.CustomUserDetailsService;
 
 import java.util.stream.Collectors;
 
@@ -18,6 +21,7 @@ import java.util.stream.Collectors;
 public class IngredientController {
     private final IngredientRepository repo;
     private final IngredientMapper ingredientMapper;
+    private final CustomUserDetailsService customUserDetailsService;
 
     @GetMapping
     public Iterable<IngredientDto> allIngredients() {
@@ -27,16 +31,21 @@ public class IngredientController {
     }
 
     @PostMapping
-    @Secured("ROLE_ADMIN")
+//    @PreAuthorize("hasAuthority('ROLE_ADMIN')")
     @ResponseStatus(HttpStatus.CREATED)
     public Ingredient saveIngredient(@RequestBody IngredientDto ingredient) {
         return repo.save(ingredientMapper.ingredientDtoToIngredient(ingredient));
     }
 
     @DeleteMapping("/{id}")
-    @Secured("ROLE_ADMIN")
+//    @PreAuthorize("hasAuthority('ADMIN')")
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void deleteIngredient(@PathVariable("id") String ingredientId) {
+    public ResponseEntity<?> deleteIngredient(@PathVariable("id") String ingredientId, @RequestBody SignUpDto signUpDto) {
+        UserDetails userDetails = customUserDetailsService.loadUserByUsername(signUpDto.getLogin());
+        if(userDetails == null){
+            return new ResponseEntity<>("Username is already taken!", HttpStatus.BAD_REQUEST);
+        }
         repo.deleteById(ingredientId);
+        return new ResponseEntity<>(null, HttpStatus.OK);
     }
 }
